refactor: Создает отдельные папки-направления для shared/lib
