+++ src/shared/lib/misc/isFalsy.ts
