+++ src/shared/lib/misc/isTruthy.ts
